Migrate some structures to zerocopy

+++ src/signatures/shrs.rs
@@ -22,7 +22,7 @@ pub fn shrs_parser(file_data: &[u8], offset: usize) -> Result<SignatureResult, S
     };
 
     if let Ok(shrs_header) = parse_shrs_header(&file_data[offset..]) {
-        result.size = shrs_header.header_size + shrs_header.data_size;
+        result.size = shrs_header.header_size + shrs_header.data_size as usize;
         result.description = format!(
             "{}, header size: {} bytes, encrypted data size: {} bytes, IV: {}",
             result.description,
